[Refactor] Split the model.py file into multiple modules, one for each class, located in the subpackage eulerpi.core.models. Refactor usage of ArtificialModelInterface. Add forward_vectorized to (Base)Model and JaxModel.
