--- docs/source/eulerpi.core.model.rst
@@ -1,7 +1,0 @@
-eulerpi.core.model module
-=========================
-
-.. automodule:: eulerpi.core.model
-   :members:
-   :undoc-members:
-   :show-inheritance:
