+++ docs/source/eulerpi.core.models.artificial_model_interface.rst
@@ -1,0 +1,7 @@
+eulerpi.core.models.artificial\_model\_interface module
+=======================================================
+
+.. automodule:: eulerpi.core.models.artificial_model_interface
+   :members:
+   :undoc-members:
+   :show-inheritance:
