+++ docs/source/eulerpi.core.models.base_model.rst
@@ -1,0 +1,7 @@
+eulerpi.core.models.base_model module
+================================
+
+.. automodule:: eulerpi.core.models.base_model
+   :members:
+   :undoc-members:
+   :show-inheritance:
