+++ docs/source/eulerpi.core.models.jax_extension.rst
@@ -1,0 +1,7 @@
+eulerpi.core.models.jax\_extension module
+=========================================
+
+.. automodule:: eulerpi.core.models.jax_extension
+   :members:
+   :undoc-members:
+   :show-inheritance:
