+++ docs/source/eulerpi.core.models.jax_model.rst
@@ -1,0 +1,7 @@
+eulerpi.core.models.jax\_model module
+=====================================
+
+.. automodule:: eulerpi.core.models.jax_model
+   :members:
+   :undoc-members:
+   :show-inheritance:
