+++ docs/source/eulerpi.core.models.rst
@@ -1,0 +1,11 @@
+eulerpi.core.models package
+===========================
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Classes
+
+.. automodule:: eulerpi.core.models
+   :members:
+   :undoc-members:
+   :show-inheritance:
