+++ docs/source/eulerpi.core.models.sbml_model.rst
@@ -1,0 +1,7 @@
+eulerpi.core.models.sbml\_model module
+======================================
+
+.. automodule:: eulerpi.core.models.sbml_model
+   :members:
+   :undoc-members:
+   :show-inheritance:
