+++ docs/source/eulerpi.examples.cpp.cpp_model.rst
@@ -1,0 +1,7 @@
+eulerpi.examples.cpp.cpp\_model module
+======================================
+
+.. automodule:: eulerpi.examples.cpp.cpp_model
+   :members:
+   :undoc-members:
+   :show-inheritance:
