--- docs/source/eulerpi.jax_extension.rst
@@ -1,7 +1,0 @@
-eulerpi.jax\_extension module
-=============================
-
-.. automodule:: eulerpi.jax_extension
-   :members:
-   :undoc-members:
-   :show-inheritance:
