+++ docs/source/eulerpi.rst
@@ -1,26 +1,8 @@
 eulerpi package
 ===============
 
-Subpackages
------------
-
 .. toctree::
    :maxdepth: 4
 
    Core<eulerpi.core>
-
-Submodules
-----------
-
-.. toctree::
-   :maxdepth: 4
-
-   Jax Extension<eulerpi.jax_extension>
-
-Module contents
----------------
-
-.. automodule:: eulerpi
-   :members:
-   :undoc-members:
-   :show-inheritance:
+   Examples<eulerpi.examples>
