+++ eulerpi/examples/corona/corona.py
@@ -5,7 +5,7 @@ import jax.numpy as jnp
 import numpy as np
 
 from eulerpi import logger
-from eulerpi.core.model import ArtificialModelInterface, JaxModel
+from eulerpi.core.models import ArtificialModelInterface, JaxModel
 
 
 class Corona(JaxModel):
@@ -30,9 +30,9 @@ class Corona(JaxModel):
     .. note::
 
         * ODE Solver: To solve the ODE problem the jax based ode solver library :code:`diffrax` is used: https://github.com/patrick-kidger/diffrax.
-        * Automatic Differentiation: The derivatives are calculated automatically with jax by deriving from the class :py:class:`~eulerpi.core.model.JaxModel`,
-          which automatically calculates sets :py:meth:`~eulerpi.core.model.Model.jacobian`.
-        * JIT compilation: Inheriting from :py:class:`~eulerpi.core.model.JaxModel` also enables jit compilation / optimization for the forward and jacobian method.
+        * Automatic Differentiation: The derivatives are calculated automatically with jax by deriving from the class :py:class:`~eulerpi.core.models.JaxModel`,
+          which automatically calculates sets :py:meth:`~eulerpi.core.models.BaseModel.jacobian`.
+        * JIT compilation: Inheriting from :py:class:`~eulerpi.core.models.JaxModel` also enables jit compilation / optimization for the forward and jacobian method.
           This usually results in a significant execution speedup. It also allows to run your model on the gpu.
     """
